no compile error in boxes.cpp
no compile error in pdflib_core.cpp

It needs to be checked for its functionality if it is correct or not.

no compile errors in pslib.cpp but there are for sure some logical errors need to be checked

No compile errors in scpageoutput

No compile errors in scpageoutput but for sure there are logical errors
such as shrink() which need to be re-implemented

No compile error in scribusview

No compile error in sctextstruct

Correction for all previous commit

There were some stupid mistak in previous commits and I modified it now

No compile error in svgexplugin

There is no compile error in whole project

we play around some compile error to make the code compiled first but it
has some logical problems.
The main logic error that we need to fix is in using linecontrol  struct

No compile error

No compile error in the code. The code was written by Vox is fixed from
syntax side but now it need to be complete.

fix pull conflict

fixing continous integration