Implementation of render method
Render method (2 bug)

render method work nicely for screen with only 2 small bug. first one
the text get bolder when we ever we write more text. the other bug is
the text dosent stop at left edge of the frame as it suppose to do but
it goes outside the frame

bold issue in last commit

rendor method complete