Fix assumption of making glyphBox representing one char
I put the ground to fix this problem but I did not modify all codes
accordingly because I did not Know all the cases that we make this
assumption. What I did is that I add firstChar and lastChar in glyphrun
that will indicate how many chars in this run and for now I set
firstChar = lastChar = index of Char in story text. I need to modify any
function that send index of the char to pageitem to make it send also
first and last char index of the run if we have more than one char for
that run.