Speed TextFramePainter a bit
That is so much we can do here to speed glyph rendering, most of time
now is spent in QPainter::drawPixmap(), and too much lesser extent in
cairo_show_glyphs().

Text selection still lags a bit, refactoring the code to just draw the
selection instead of the whole frame needs more investigation.