fix selectWord to be aware of CTL
selectWord is used to select word when you double click on a text frame. Although, it works fine with RTL,
it fails to work with CJK, Thai, and east Asia languages. I tried to use QTextBoundaryFinder but it has
bugs with CJK so I move to ICU which it works fine.
This part of #166