shift back to UTF-16
It has been discussed a lot in the mailing list whether we shift to
UFT-32 or not, see http://lists.scribus.net/pipermail/scribus-
dev/2016-June/003121.html

As well as,moving to UTF-32 causes some bugs in storytext class because
ICU use UTF-16 and it needs a bit of work to make it use UTF-32.
For now I revert all commits related to merging to UTF-32 and create a
new branch if we want them in future.