add Search \ replace all text frames support
Here is a patch will implement this feature. The work is done by @Majda.

The current status now:
- every thing is working fine, you can search and replace into all text
frames.
- the logic behind search for all text frames is explained as follow:
The search algorithm is to start from selected text frame, if there is
no
text frame selected select the first text frame in the current page, if
not, find the first text frame from next pages, if reaching the end of
the document start searching from the first page.

Enhancement may be added in the future:
- make search and replace dialog is modeless, so you can search and
interact with text directly.